requirements.txt
apache-airflow==2.7.2
certifi==2025.7.14
charset-normalizer==3.4.2
docopt==0.6.2
idna==3.10
pipreqs==0.4.13
requests==2.32.4
urllib3==2.5.0
yarg==0.1.10
pandas
